Multiple fixes (#31)
* Implement scard and srem

* fix exists command implementation

* respect TTL in "exists"

* fix `sadd` implementation

fixes a bug in `sadd` implementation: this operation should return
the number of added entries, not the size of resulting set

* fix syntax error

* fix `delete` implementation

* implement `time`

* test scope for embedded-redis dependency

+++ src/main/java/com/github/fppt/jedismock/operations/RO_del.java
@@ -11,12 +11,11 @@ class RO_del extends AbstractRedisOperation {
         super(base, params);
     }
 
-    Slice response(){
+    Slice response() {
         int count = 0;
         for (Slice key : params()) {
-            Slice value = base().getValue(key);
-            base().deleteValue(key);
-            if (value != null) {
+            if (base().exists(key)) {
+                base().deleteValue(key);
                 count++;
             }
         }
